+++ packages/vibrant-core/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.80.0](https://github.com/pedaling/opensource/compare/vibrant-core-0.79.2...vibrant-core-0.80.0) (2023-09-21)
+
+### [0.79.2](https://github.com/pedaling/opensource/compare/vibrant-core-0.80.0-0...vibrant-core-0.79.2) (2023-09-21)
+
+## [0.80.0-0](https://github.com/pedaling/opensource/compare/vibrant-core-0.79.1...vibrant-core-0.80.0-0) (2023-09-21)
+
+### [0.79.1](https://github.com/pedaling/opensource/compare/vibrant-core-0.79.0...vibrant-core-0.79.1) (2023-09-18)
+
+
+### Features
+
+* add writingMode prop, and display web_block ([#856](https://github.com/pedaling/opensource/issues/856)) ([027f400](https://github.com/pedaling/opensource/commit/027f400e0da1b89433f2039f4a32603ad9f86d2d))
+
+## [0.79.0](https://github.com/pedaling/opensource/compare/vibrant-core-0.78.1...vibrant-core-0.79.0) (2023-09-11)
+
+### [0.78.1](https://github.com/pedaling/opensource/compare/vibrant-core-0.78.0...vibrant-core-0.78.1) (2023-09-11)
+
+## [0.76.0](https://github.com/pedaling/opensource/compare/vibrant-core-0.75.2...vibrant-core-0.76.0) (2023-09-06)
+
 ### [0.75.2](https://github.com/pedaling/opensource/compare/vibrant-core-0.75.1...vibrant-core-0.75.2) (2023-08-28)
 
 ### [0.75.1](https://github.com/pedaling/opensource/compare/vibrant-core-0.75.0...vibrant-core-0.75.1) (2023-08-25)